ficed theme spaceunit error with *2
